// web/client/src/app/tshepii/closest-pair.ts
/* 
* NOTICE
*  
* This software (or technical data) was produced for the U. S. Government
* and is subject to the Rights in Data-General Clause 52.227-14, Alt. IV
* (May 2014) – Alternative IV (Dec 2007)
* 
* (c) 2024 The MITRE Corporation. All Rights Reserved.
*/ 

import Point = paper.Point;
import { Annotation } from '../types/annotation';

export class ClosestPair {
  distance: number;
  point1: Point;
  point2: Point;
  anno1: Annotation;
  anno2: Annotation;
}